removing logs, rename branch.js to treeNode.js, also renamed the import
